Move frame_manager to include folder
