+++ rviz_common/CMakeLists.txt
@@ -86,7 +86,7 @@ set(rviz_common_headers_to_moc
   include/rviz_common/display_context.hpp
   # src/rviz_common/displays_panel.hpp
   src/rviz_common/failed_panel.hpp
-  src/rviz_common/frame_manager.hpp
+  include/rviz_common/frame_manager.hpp
   src/rviz_common/loading_dialog.hpp
   src/rviz_common/new_object_dialog.hpp
   src/rviz_common/panel.hpp
