+++ rviz_common/src/rviz_common/frame_manager.cpp
@@ -28,7 +28,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "./frame_manager.hpp"
+#include "rviz_common/frame_manager.hpp"
 
 #include <algorithm>
 #include <chrono>
