+++ rviz_common/src/rviz_common/frame_position_tracking_view_controller.cpp
@@ -44,7 +44,7 @@
 #endif
 
 #include "rviz_common/display_context.hpp"
-#include "./frame_manager.hpp"
+#include "rviz_common/frame_manager.hpp"
 #include "rviz_common/properties/enum_property.hpp"
 #include "./properties/tf_frame_property.hpp"
 #include "./view_manager.hpp"
