+++ rviz_common/src/rviz_common/properties/tf_frame_property.cpp
@@ -36,7 +36,7 @@
 
 #include "tf2_ros/transform_listener.h"
 
-#include "../frame_manager.hpp"
+#include "rviz_common/frame_manager.hpp"
 
 namespace rviz_common
 {
