+++ rviz_common/src/rviz_common/temp/default_plugins/displays/grid_display.cpp
@@ -47,7 +47,7 @@
 #endif
 
 #include "rviz_common/display_context.hpp"
-#include "../../../frame_manager.hpp"
+#include "rviz_common/frame_manager.hpp"
 #include "rviz_rendering/grid.hpp"
 #include "../../../properties/parse_color.hpp"
 #include "rviz_common/properties/property.hpp"
