+++ rviz_common/src/rviz_common/time_panel.cpp
@@ -37,7 +37,7 @@
 #include <QComboBox>
 
 #include "./visualization_manager.hpp"
-#include "./frame_manager.hpp"
+#include "rviz_common/frame_manager.hpp"
 
 #include "./display_group.hpp"
